==> notebooks/6-visualize-sound-wave.ipynb <==
%matplotlib inline
%reload_ext autoreload
%autoreload 2
%config InlineBackend.figure_format = 'retina'

import numpy as np
from scipy.io import wavfile
import matplotlib.pyplot as plt

# https://pythontic.com/visualization/signals/spectrogram
samplingFrequency, signalData = wavfile.read('data/voice.wav')

# # Repeat the silence in the beginning
# n = 5000
# signalData_padded = np.concatenate((np.tile(signalData[:n], (len(signalData) // n + 5, 1)), signalData))
# signalData_padded.shape

# # Remove the silence in the end
# n2 = 40000
# signalData_padded = signalData_padded[:-n2]
# signalData_padded.shape

# Plot
fig = plt.figure(figsize=(20, 5))
plt.plot(signalData[:-15000], c='#1f77b4', alpha=.2)
plt.axis('off')

fig.savefig('data/voice.png', transparent=True, bbox_inches='tight')
